+++ app/Main.hs
@@ -35,8 +35,12 @@ import Concordium.Types.Queries as Types
 import Data.Range.Parser
 import Options.Applicative
 
+serverHost :: String
+serverHost = "0.0.0.0"
+
 data ProxyConfig = ProxyConfig
     { pcGRPC :: GrpcConfig,
+      pcPort :: Int,
       pcDBConnString :: ByteString,
       -- | Account file used for GTU drop. Only used on stagenet and testnet.
       pcGTUAccountFile :: Maybe FilePath,
@@ -63,6 +67,7 @@ parser =
         mkProxyConfig
             <$> backendParser
             <*> optional (option auto (long "grpc-timeout" <> value 15 <> showDefault <> metavar "TIMEOUT" <> help "Timeout of grpc requests."))
+            <*> option auto (long "port" <> metavar "PORT" <> help "Port number to run the server on" <> showDefault <> value 3000)
             <*> strOption (long "db" <> metavar "STR" <> help "database connection string")
             <*> optional (strOption (long "drop-account" <> metavar "FILE" <> help "file with CCD drop account credentials (only used for stagenet and testnet)."))
             <*> option (eitherReader amountFromStringInform) (long "drop-amount" <> metavar "CCD-AMOUNT" <> help "Amount of CCDs to drop upon request." <> value 2_000_000_000)
@@ -86,13 +91,13 @@ parser =
                 (Just (fromMaybe 15 timeout))
                 (CMDS.grpcUseTls backend)
 
-runSite :: (YesodDispatch site) => Int -> Network.Wai.Handler.Warp.HostPreference -> site -> IO ()
-runSite port host site = do
+runSite :: (YesodDispatch site) => Int -> site -> IO ()
+runSite port site = do
     toWaiApp site
         >>= Network.Wai.Handler.Warp.runSettings
             ( Network.Wai.Handler.Warp.setPort port $
                 Network.Wai.Handler.Warp.setServerName "Concordium-wallet-proxy" $
-                    Network.Wai.Handler.Warp.setHost host $
+                    Network.Wai.Handler.Warp.setHost (fromString serverHost) $
                         Network.Wai.Handler.Warp.defaultSettings
             )
 
@@ -177,6 +182,7 @@ main = do
         $logDebug ("Using iOS V1 update config: " <> fromString (show forcedUpdateConfigIOSV1))
         $logDebug ("Using Android V1 update config: " <> fromString (show forcedUpdateConfigAndroidV1))
         $logDebug ("Using logLevel: " <> fromString (show logLevel))
+        $logDebug ("Running server on: " <> fromString serverHost <> ":" <> fromString (show pcPort))
     runStderrLoggingT . filterL $ withPostgresqlPool pcDBConnString 10 $ \dbConnectionPool -> liftIO $ do
         -- do not care about the gtu receipients database if gtu drop is not enabled
         when (isJust gtuDropData) $ runSqlPool (runMigration migrateGTURecipient) dbConnectionPool
@@ -190,5 +196,5 @@ main = do
                     Right res -> case getResponseValue res of
                         Right cParams -> do
                             let globalInfo = toJSON $ Versioned (Version 0) cParams
-                            runSite 3000 "0.0.0.0" Proxy{grpcEnvData = cfg, ..}
+                            runSite pcPort Proxy{grpcEnvData = cfg, ..}
                         Left (_, err) -> die $ "Cannot obtain cryptographic parameters due to error: " ++ err